module refactoring (#18)
